Standard GOV UK icons added to toolkit
Previously icons were held in a seperate static repository which would need to be hosted in addition to the frontend toolkit. Putting the icons into this toolkit helps to get all of the resources into one place. It also makes it easier for some developers to pick up any changes to the icons using the NPM/GEM versions.