Add browser support documentation to CSS 3 mixins
Data comes from http://caniuse.com/.

I've gone a bit over the top in adding documentation for this. I'm not sure
if this is too far, but I think it's useful. I'm trying to avoid being
prescriptive, but we should be trying to drop old prefixes and not leave
them hanging around forever.

This commit changes the order of the prefixes for gradient() to make them
line up neatly, but should cause no user facing changes.

It also removes the Microsoft prefix on the transition property, because
that has never been a thing.