Unify analytics interfaces
So that in a future commit I can make Tracker.js work when it is given
only one of universalId or classicId (which is needed for turning off
classic analytics), make the interfaces to
google-analytics-classic-tracker and google-analytics-universal-tracker
the same.

This means we can treat them both as generic tracker objects in
tracker.js and tracker.js can make the same calls without needing to
know what type of tracker it is.