Tag the repo when version is changed
When we make a change to the version file we automatically create and
publish new versions of the toolkit packages.

This change will then also create a tag on this repository with the new
version number.

This has two benefits. It will be much easier to view the differences
between versions as we can just look at the changes between friendly tag
names. The other benefit is things like bower can pull in a specific tag
rather than a sha.