Provide multivariate test cohort to analytics dimension
Google's new universal analytics setDimension call takes 2 arguments: the
dimension index and the custom data to set for this session.  Our
MultivariateTest still sends 4 arguments: the dimension index, the name of
the test, the name of the cohort, and the scope in which to set these
values.  This set of arguments was the pattern for the old "classic"
analytics component, but the new pattern discards the extra 2 arguments.
This means we don't get any details of the cohort sent to the tracker.

We don't need the 4th argument any more so we can discard that. We do want
the cohort though, so we combine it with the name of the test and send that
as the custom data for the dimension.