Add GOVUKTracker module
This has a similar API to the GA tracker module so it’s compatible with
the `GOVUK.Analytics` module. It sends data to the server by adding an
`img` tag for an empty gif with the payload data encoded in the
querystring of the `src`. This data will be picked up from the CDN logs.

It adds some additional data that the main GA tracker doesn’t add, such
as performance information and the GA client ID. This is because it’s
assumed that GA sends similar things by default anyway.

This data will be used alongside regular exports of content metadata to
provide performance analysts with more reliable content-focused
analytics data.