Disable tracking in GA
I'm finding this difficult to test but this does seem to stop requests
to ga-audience and doubleclick.

Discovered when checking some waterfall logs, seen in external tools such as
SpeedCurve, WebPageTest.

To test I tried this locally, it seems these endpoints are only called on invitial visit
which makes it hard to reproduce.